i2c: Add a BCM283x slave I2C device
There is a fairly limited I2C slave device on the BCM283x and family, including
the BCM2711.  The device is fairly broken, but it is capable of being an IPMI
IPMB device, so you could use a Raspberry Pi (at least a Pi 4, other Pis may or
may not run the signals to the header) as an IPMI BMC, or as a management
controller on an IPMB bus.

Signed-off-by: Corey Minyard <minyard@acm.org>